第一回合勉强成功

+++ Core/Src/zigbee.c
@@ -1,23 +1,22 @@
-#include"zigbee.h"
+#include "zigbee.h"
 #include "string.h"
 #include "value.h"
 #include "main.h"
 
-uint8_t zigbeeMessage[ZIGBEE_MESSAGE_LENTH];   //经过校验后确认无误的信息
-uint8_t zigbeeReceive[zigbeeReceiveLength];   //实时记录收到的信息
-uint8_t zigbeeMessageSend[4] = {0x00,0x01,0x02,0x03};  //小车给上位机发送的信息
-uint8_t zigbee_halfrecv_buff[ZIGBEE_MESSAGE_LENTH];    //储存未全部被接收的消息
-UART_HandleTypeDef* zigbee_huart;
-uint8_t zigbee_halfrecv_len = 0;                      //储存未完全接收消息的长度
-
+uint8_t zigbeeMessage[ZIGBEE_MESSAGE_LENTH];			 //经过校验后确认无误的信息
+uint8_t zigbeeReceive[zigbeeReceiveLength];				 //实时记录收到的信息
+uint8_t zigbeeMessageSend[4] = {0x00, 0x01, 0x02, 0x03}; //小车给上位机发送的信息
+uint8_t zigbee_halfrecv_buff[ZIGBEE_MESSAGE_LENTH];		 //储存未全部被接收的消息
+UART_HandleTypeDef *zigbee_huart;
+uint8_t zigbee_halfrecv_len = 0; //储存未完全接收消息的长度
 
 /***********************接口****************************/
 
 void zigbee_Init(UART_HandleTypeDef *huart)
 {
 	zigbee_huart = huart;
-	__HAL_UART_ENABLE_IT(zigbee_huart, UART_IT_IDLE);		//开启空闲中断
-	HAL_UART_Receive_DMA(zigbee_huart, zigbeeReceive, zigbeeReceiveLength);    //开启DMA中断接收
+	__HAL_UART_ENABLE_IT(zigbee_huart, UART_IT_IDLE);						//开启空闲中断
+	HAL_UART_Receive_DMA(zigbee_huart, zigbeeReceive, zigbeeReceiveLength); //开启DMA中断接收
 }
 
 void zigbeeMessageRecord(uint8_t rx_length)
@@ -26,7 +25,7 @@ void zigbeeMessageRecord(uint8_t rx_length)
 	{
 		memcpy(zigbeeMessage, zigbeeReceive, ZIGBEE_MESSAGE_LENTH);
 	}
-	else if (rx_length < ZIGBEE_MESSAGE_LENTH)   //用于处理单条消息分两次发送的情况
+	else if (rx_length < ZIGBEE_MESSAGE_LENTH) //用于处理单条消息分两次发送的情况
 	{
 		if (!zigbee_halfrecv_len)
 		{
@@ -36,8 +35,7 @@ void zigbeeMessageRecord(uint8_t rx_length)
 		}
 		else if (zigbee_halfrecv_len && rx_length + zigbee_halfrecv_len == ZIGBEE_MESSAGE_LENTH)
 		{
-			if ((zigbeeReceive[rx_length - 2] == 0x0D && zigbeeReceive[rx_length - 1] == 0x0A)
-			 || (rx_length == 1 && zigbeeReceive[0] == 0x0A && zigbee_halfrecv_buff[ZIGBEE_MESSAGE_LENTH - 2] == 0x0D))
+			if ((zigbeeReceive[rx_length - 2] == 0x0D && zigbeeReceive[rx_length - 1] == 0x0A) || (rx_length == 1 && zigbeeReceive[0] == 0x0A && zigbee_halfrecv_buff[ZIGBEE_MESSAGE_LENTH - 2] == 0x0D))
 			{
 				memcpy(zigbeeMessage, zigbee_halfrecv_buff, zigbee_halfrecv_len);
 				memcpy(zigbeeMessage + zigbee_halfrecv_len, zigbeeReceive, rx_length);
@@ -47,7 +45,7 @@ void zigbeeMessageRecord(uint8_t rx_length)
 
 			zigbee_halfrecv_len = 0;
 		}
-		else 
+		else
 		{
 			zigbee_halfrecv_len = 0;
 			return;
@@ -56,18 +54,18 @@ void zigbeeMessageRecord(uint8_t rx_length)
 	else
 		return;
 
-	//HAL_GPIO_TogglePin(pinLED_GPIO_Port, pinLED_Pin);
+	// HAL_GPIO_TogglePin(pinLED_GPIO_Port, pinLED_Pin);
 	return;
 }
 
 void zigbeeSend(int MineType)
 {
-	HAL_UART_Transmit(zigbee_huart,zigbeeMessageSend+MineType,1,HAL_MAX_DELAY);
+	HAL_UART_Transmit(zigbee_huart, zigbeeMessageSend + MineType, 1, HAL_MAX_DELAY);
 }
 
-int32_t getGameTime(void) 
+int32_t getGameTime(void)
 {
-	return (int32_t)((zigbeeMessage[0] << 8) + zigbeeMessage[1]);
+	return (int32_t)(((uint16_t)zigbeeMessage[0] << 8) + (uint16_t)zigbeeMessage[1]);
 }
 
 int32_t getGameState(void)
@@ -77,21 +75,22 @@ int32_t getGameState(void)
 
 int32_t getCarTask(void)
 {
-	//return 0;
+	// return 0;
 	return (int32_t)((zigbeeMessage[2] & 0x20) >> 5);
 }
 
-int32_t getIsMineIntensityValid(int MineNo) {
-	if (MineNo == 0 )
+int32_t getIsMineIntensityValid(int MineNo)
+{
+	if (MineNo == 0)
 	{
-		return (int32_t)((zigbeeMessage[2] &0x10) >>4 );
+		return (int32_t)((zigbeeMessage[2] & 0x10) >> 4);
 	}
-	else if(MineNo == 1)
+	else if (MineNo == 1)
 	{
-		return (int32_t)((zigbeeMessage[2] &0x08) >>3);
+		return (int32_t)((zigbeeMessage[2] & 0x08) >> 3);
 	}
 	else
-	{	
+	{
 		return (int32_t)INVALID_ARG;
 	}
 }
@@ -106,33 +105,33 @@ int32_t getCarPosY()
 	return (int32_t)(uint16_t)(zigbeeMessage[4]);
 }
 
-int32_t getMineArrayType(int MineNo) {
+int32_t getMineArrayType(int MineNo)
+{
 	if (MineNo == 0)
 	{
 		return (int32_t)(uint16_t)((zigbeeMessage[5] & 0xC0) >> 6);
-	
 	}
-	else if(MineNo == 1)
-	{	
+	else if (MineNo == 1)
+	{
 		return (int32_t)(uint16_t)((zigbeeMessage[5] & 0x30) >> 4);
 	}
-	else 
+	else
 	{
 		return (int32_t)(uint16_t)INVALID_ARG;
 	}
 }
 
-int32_t getCarMineSumNum(void) 
+int32_t getCarMineSumNum(void)
 {
 	return (int32_t)(uint16_t)(zigbeeMessage[5] & 0x0F);
 }
 
-int32_t getCarMineANum(void) 
+int32_t getCarMineANum(void)
 {
 	return (int32_t)(uint16_t)((zigbeeMessage[6] & 0xF0) >> 4);
 }
 
-int32_t getCarMineBNum(void) 
+int32_t getCarMineBNum(void)
 {
 	return (int32_t)(uint16_t)(zigbeeMessage[6] & 0x0F);
 }
@@ -142,26 +141,26 @@ int32_t getCarMineCNum(void)
 	return (int32_t)(uint16_t)((zigbeeMessage[7] & 0xF0) >> 4);
 }
 
-int32_t getCarMineDNum(void) 
+int32_t getCarMineDNum(void)
 {
 	return (int32_t)(uint16_t)(zigbeeMessage[7] & 0x0F);
 }
 
 int32_t getMyBeaconMineType(int MineType)
 {
-	if (MineType == 0 )
-	{	
+	if (MineType == 0)
+	{
 		return (int32_t)(uint16_t)((zigbeeMessage[8] & 0xC0) >> 6);
 	}
-	else if(MineType == 1)
+	else if (MineType == 1)
 	{
 		return (int32_t)(uint16_t)((zigbeeMessage[8] & 0x30) >> 4);
 	}
-	else if(MineType == 2)
+	else if (MineType == 2)
 	{
 		return (int32_t)(uint16_t)((zigbeeMessage[8] & 0x0C) >> 2);
 	}
-	else 
+	else
 	{
 		return (int32_t)(uint16_t)INVALID_ARG;
 	}
@@ -171,101 +170,125 @@ int32_t getMineIntensity(int MineNo)
 {
 	if (MineNo == 0 || MineNo == 1)
 	{
-		return (int32_t)(uint32_t)((zigbeeMessage[9+4*MineNo] << 24) + (zigbeeMessage[10 + 4 * MineNo] << 16) + (zigbeeMessage[11 + 4 * MineNo] << 8) + (zigbeeMessage[12 + 4 * MineNo]));
+		// return (int32_t)(uint32_t)0xffff;
+		return (int32_t)((((uint32_t)zigbeeMessage[9 + 4 * MineNo]) << 24) + (((uint32_t)zigbeeMessage[10 + 4 * MineNo]) << 16) + (((uint32_t)zigbeeMessage[11 + 4 * MineNo]) << 8) + ((uint32_t)zigbeeMessage[12 + 4 * MineNo]));
 	}
 	else
 	{
 		return (int32_t)(uint32_t)INVALID_ARG;
-	}	
+	}
 }
 
-int32_t getDistanceOfMyBeacon(int BeaconNo) {
+int32_t getDistanceOfMyBeacon(int BeaconNo)
+{
 	if (BeaconNo != 0 && BeaconNo != 1 && BeaconNo != 2)
 		return (int32_t)(uint16_t)INVALID_ARG;
 	else
-		return (int32_t)(uint16_t)((zigbeeMessage[17+2*BeaconNo] << 8) + zigbeeMessage[18+2*BeaconNo]);
+		return (int32_t)(uint16_t)(((uint16_t)zigbeeMessage[17 + 2 * BeaconNo] << 8) + (uint16_t)zigbeeMessage[18 + 2 * BeaconNo]);
 }
 
-int32_t getDistanceOfRivalBeacon(int BeaconNo) {
+int32_t getDistanceOfRivalBeacon(int BeaconNo)
+{
 	if (BeaconNo != 0 && BeaconNo != 1 && BeaconNo != 2)
 		return (int32_t)(uint16_t)INVALID_ARG;
 	else
-		return (int32_t)(uint16_t)((zigbeeMessage[23+2*BeaconNo] << 8) + zigbeeMessage[24+2*BeaconNo]);
+		return (int32_t)(uint16_t)(((uint16_t)zigbeeMessage[23 + 2 * BeaconNo] << 8) + (uint16_t)zigbeeMessage[24 + 2 * BeaconNo]);
 }
 
-int32_t getParkDotMineType(int ParkDotNo) 
+int32_t getParkDotMineType(int ParkDotNo)
 {
 	if (ParkDotNo < 0 || ParkDotNo > 7)
 		return (int32_t)(uint16_t)INVALID_ARG;
 	else
 		switch (ParkDotNo)
 		{
-		case 0: return (int32_t)(uint16_t)((zigbeeMessage[29] & 0xC0) >> 6);break;
-	    case 1: return (int32_t)(uint16_t)((zigbeeMessage[29] & 0x30) >> 4);break;
-		case 2: return (int32_t)(uint16_t)((zigbeeMessage[29] & 0x0C) >> 2);break;
-		case 3: return (int32_t)(uint16_t)((zigbeeMessage[29] & 0x03));break;
-		case 4: return (int32_t)(uint16_t)((zigbeeMessage[30] & 0xC0) >> 6);break;
-		case 5: return (int32_t)(uint16_t)((zigbeeMessage[30] & 0x30) >> 4);break;
-		case 6: return (int32_t)(uint16_t)((zigbeeMessage[30] & 0x0C) >> 2);break;
-		case 7: return (int32_t)(uint16_t)((zigbeeMessage[30] & 0x03));break;
+		case 0:
+			return (int32_t)(uint16_t)((zigbeeMessage[29] & 0xC0) >> 6);
+			break;
+		case 1:
+			return (int32_t)(uint16_t)((zigbeeMessage[29] & 0x30) >> 4);
+			break;
+		case 2:
+			return (int32_t)(uint16_t)((zigbeeMessage[29] & 0x0C) >> 2);
+			break;
+		case 3:
+			return (int32_t)(uint16_t)((zigbeeMessage[29] & 0x03));
+			break;
+		case 4:
+			return (int32_t)(uint16_t)((zigbeeMessage[30] & 0xC0) >> 6);
+			break;
+		case 5:
+			return (int32_t)(uint16_t)((zigbeeMessage[30] & 0x30) >> 4);
+			break;
+		case 6:
+			return (int32_t)(uint16_t)((zigbeeMessage[30] & 0x0C) >> 2);
+			break;
+		case 7:
+			return (int32_t)(uint16_t)((zigbeeMessage[30] & 0x03));
+			break;
 		default:
 			return (int32_t)(uint16_t)((zigbeeMessage[30] & 0x03));
 		}
 }
 
-int32_t getCarZone(void) {
+int32_t getCarZone(void)
+{
 	return (int32_t)(uint16_t)((zigbeeMessage[31] & 0x80) >> 7);
 }
 
-int32_t getIsCarPosValid(void) {
+int32_t getIsCarPosValid(void)
+{
 	return (int32_t)(uint16_t)((zigbeeMessage[31] & 0x40) >> 6);
 }
 
-int32_t getIsDistanceOfMyBeaconValid(int BeaconNo) {
-	if (BeaconNo == 0 )
+int32_t getIsDistanceOfMyBeaconValid(int BeaconNo)
+{
+	if (BeaconNo == 0)
 	{
-		return (int32_t)(uint16_t)((zigbeeMessage[31]& 0x20) >> 5);
+		return (int32_t)(uint16_t)((zigbeeMessage[31] & 0x20) >> 5);
 	}
-	else if(BeaconNo == 1)
+	else if (BeaconNo == 1)
 	{
 		return (int32_t)(uint16_t)((zigbeeMessage[31] & 0x10) >> 4);
 	}
-	else if(BeaconNo == 2)
+	else if (BeaconNo == 2)
 	{
 		return (int32_t)(uint16_t)((zigbeeMessage[31] & 0x08) >> 3);
 	}
-	else 
+	else
 	{
 		return (int32_t)(uint16_t)INVALID_ARG;
-	}		
+	}
 }
 
-int32_t getIsDistanceOfRivalBeaconValid(int BeaconNo) {
-	if (BeaconNo == 0 )
+int32_t getIsDistanceOfRivalBeaconValid(int BeaconNo)
+{
+	if (BeaconNo == 0)
 	{
-		return (int32_t)(uint16_t)((zigbeeMessage[31]& 0x04) >> 2);
+		return (int32_t)(uint16_t)((zigbeeMessage[31] & 0x04) >> 2);
 	}
-	else if(BeaconNo == 1)
+	else if (BeaconNo == 1)
 	{
 		return (int32_t)(uint16_t)((zigbeeMessage[31] & 0x02) >> 1);
 	}
-	else if(BeaconNo == 2)
+	else if (BeaconNo == 2)
 	{
 		return (int32_t)(uint16_t)((zigbeeMessage[31] & 0x01));
 	}
-	else 
+	else
 	{
 		return (int32_t)(uint16_t)INVALID_ARG;
 	}
 }
 
-int32_t getCarScore(void) {
-	return (int32_t)(int16_t)((int8_t)(zigbeeMessage[32]<<8) + zigbeeMessage[33] );
+int32_t getCarScore(void)
+{
+	return (int32_t)(int16_t)(((int16_t)zigbeeMessage[32] << 8) + (int16_t)zigbeeMessage[33]);
 }
 
 int32_t getMyBeaconPosX(int BeaconNo)
 {
-	if(BeaconNo !=0 && BeaconNo != 1 && BeaconNo != 2)
+	if (BeaconNo != 0 && BeaconNo != 1 && BeaconNo != 2)
 	{
 		return (int32_t)(uint16_t)INVALID_ARG;
 	}
@@ -277,7 +300,7 @@ int32_t getMyBeaconPosX(int BeaconNo)
 
 int32_t getMyBeaconPosY(int BeaconNo)
 {
-	if(BeaconNo !=0 && BeaconNo != 1 && BeaconNo != 2)
+	if (BeaconNo != 0 && BeaconNo != 1 && BeaconNo != 2)
 	{
 		return (int32_t)(uint16_t)INVALID_ARG;
 	}
@@ -289,7 +312,7 @@ int32_t getMyBeaconPosY(int BeaconNo)
 
 int32_t getRivalBeaconPosX(int BeaconNo)
 {
-	if(BeaconNo !=0 && BeaconNo != 1 && BeaconNo != 2)
+	if (BeaconNo != 0 && BeaconNo != 1 && BeaconNo != 2)
 	{
 		return (int32_t)(uint16_t)INVALID_ARG;
 	}
@@ -301,7 +324,7 @@ int32_t getRivalBeaconPosX(int BeaconNo)
 
 int32_t getRivalBeaconPosY(int BeaconNo)
 {
-	if(BeaconNo !=0 && BeaconNo != 1 && BeaconNo != 2)
+	if (BeaconNo != 0 && BeaconNo != 1 && BeaconNo != 2)
 	{
 		return (int32_t)(uint16_t)INVALID_ARG;
 	}
@@ -314,7 +337,7 @@ int32_t getRivalBeaconPosY(int BeaconNo)
 Pos getCarPos(void)
 {
 	/* test*/
-	Pos a = {(int32_t)getCarPosX(),(int32_t)getCarPosY()};
+	Pos a = {(int32_t)getCarPosX(), (int32_t)getCarPosY()};
 
 	return a;
 }
@@ -332,8 +355,3 @@ Pos getMyBeaconPos(int BeaconNo)
 }
 
 /***************************************************/
-
-
-
-
-
